+++ Source/SynHighlighterPython.pas
@@ -47,7 +47,8 @@ uses
   Vcl.Graphics,
   SynEditHighlighter,
   SynEditCodeFolding,
-  StringResources;
+  StringResources,
+  SynEditTypes;
 
 type
   TtkTokenKind = (tkComment, tkIdentifier, tkKey, tkNull, tkNumber, tkSpace,
@@ -140,6 +141,7 @@ type
       LinesToScan: TStrings; FromLine: Integer; ToLine: Integer); override;
     procedure AdjustFoldRanges(FoldRanges: TSynFoldRanges;
       LinesToScan: TStrings); override;
+    function FlowControlAtLine(Lines: TStrings; Line: Integer): TSynFlowControl; override;
     property Keywords: TStringList read FKeywords;
   published
     property CommentAttri: TSynHighlighterAttributes read fCommentAttri
@@ -633,7 +635,6 @@ procedure TSynPythonSyn.AdjustFoldRanges(FoldRanges: TSynFoldRanges;
   LinesToScan: TStrings);
 var
   I: Integer;
-  Range: TSynFoldRange;
 begin
   inherited;
   for I := 0 to FoldRanges.Count - 1 do
@@ -1327,6 +1328,45 @@ begin
         else UnknownProc;
       end;
   end;
+end;
+
+function TSynPythonSyn.FlowControlAtLine(Lines: TStrings;
+  Line: Integer): TSynFlowControl;
+var
+  SLine: string;
+  Index: Integer;
+begin
+  Result := fcNone;
+
+  SLine := Lines[Line - 1];
+
+  Index :=  SLine.IndexOf('continue');
+  if Index >= 0 then
+    Result := fcContinue
+  else
+  begin
+    Index :=  SLine.IndexOf('break');
+    if Index >= 0 then
+      Result := fcBreak
+    else
+    begin
+      Index :=  SLine.IndexOf('return');
+      if Index >= 0 then
+        Result := fcExit
+      else
+      begin
+        Index :=  SLine.IndexOf('yield');
+        if Index >= 0 then
+          Result := fcExit
+      end;
+    end;
+  end;
+
+  // Index is 0-based
+  if (Index >= 0) and
+    not (GetHighlighterAttriAtRowCol(Lines, Line - 1, Index + 1) = KeyAttri)
+  then
+    Result := fcNone;
 end;
 
 procedure TSynPythonSyn.Next;