+++ Source/frmPyIDEMain.pas
@@ -592,6 +592,9 @@
           New Features
             - Multi-caret and multi-selection editing #483
             - Annotated scrollbars
+            - Colored code structure highlight
+            - Display of program flow control symbols
+
           Issues addressed
             #1321, #1329, #1336
 
@@ -3135,7 +3138,7 @@ begin
   var TempStringList := TSmartPtr.Make(TStringList.Create)();
   if AppStorage.PathExists('Editor Options') then
   begin
-    TempStringList.AddStrings(['TrackChanges', 'SelectedColor', 'IndentGuides']);
+    TempStringList.AddStrings(['TrackChanges', 'SelectedColor', 'IndentGuides, DisplayFlowControl']);
     AppStorage.ReadPersistent('Editor Options', EditorOptions, True, True, TempStringList);
     EditorOptions.Options := EditorOptions.Options + [eoBracketsHighlight, eoAccessibility];
   end;