+++ Source/frmPyIDEMain.pas
@@ -588,7 +588,7 @@
           Issues addressed
             #1227, #1271, #1319
 
-   History:   v 5.0.2
+   History:   v 5.1.0
           New Features
             - Multi-caret and multi-selection editing #483
             - Annotated scrollbars